Created function to trim and stack qld dendro data (TrimStackQLD), created one-week plots
the function converts the 'den.qld' wide dataframe into a tall
dataframe for plotting by ggplot.

the plotting is section is laid out so as to allow the code for each
plot to be preserved.

plot names are created automatically based off the time interval that
the data is trimmed down to.

to do: put in a switch to decide if you want mm, um or mV. At the
moment it depends on how far you execute eda.R